@@ -2,7 +2,8 @@ package main
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"github.com/cloudevents/sdk-go/v2/event"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -10,15 +11,17 @@ import (
 	"strings"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestConfig(t *testing.T) {
 	os.Clearenv()
-	os.Setenv("REMINDABOT_SMTP_USER", "harry")
-	os.Setenv("REMINDABOT_SMTP_PASSWORD", "use-chats")
-	os.Setenv("REMINDABOT_SMTP_SERVER", "mock.mail.com")
-	os.Setenv("REMINDABOT_SMTP_PORT", "999")
-	os.Setenv("REMINDABOT_API_TOKEN", "123456789012")
+	_ = os.Setenv("REMINDABOT_SMTP_USER", "harry")
+	_ = os.Setenv("REMINDABOT_SMTP_PASSWORD", "use-chats")
+	_ = os.Setenv("REMINDABOT_SMTP_SERVER", "mock.mail.com")
+	_ = os.Setenv("REMINDABOT_SMTP_PORT", "999")
+	_ = os.Setenv("REMINDABOT_API_TOKEN", "123456789012")
 	c := configure()
 	if c.SmtpPort != 999 {
 		t.Errorf("expected %v, got %v", 999, c.SmtpPort)
@@ -52,6 +55,7 @@ func TestFooter(t *testing.T) {
 	if !strings.Contains(actual, rel) || !strings.Contains(actual, strconv.Itoa(year)) {
 		t.Errorf("Subject test failed expected '%s' to contain '%d' and '%s'", actual, year, rel)
 	}
+	assert.Contains(t, actual, "Pura Vida")
 }
 
 // http://www.inanzzz.com/index.php/post/fb0m/mocking-and-testing-http-clients-in-golang
@@ -64,16 +68,16 @@ func TestConsumer_GetReminders(t *testing.T) {
 		t.Error(err)
 	}
 
-	body, err := ioutil.ReadAll(resBody)
+	body, err := io.ReadAll(resBody)
 	if err != nil {
 		t.Error(err)
 	}
-	resBody.Close()
+	_ = resBody.Close()
 
 	//if http.StatusOK != res.StatusCode  {
 	//	t.Error("expected", http.StatusOK, "got", res.StatusCode)
 	//}
-	if `{"Notes":null,"EventStats":null,"ImageUrl":"","Footer":"test"}` != string(body) {
+	if `{"Notes":null,"Events":[],"ImageUrl":"","Footer":"test"}` != string(body) {
 		t.Error("expected mock server responding got", string(body))
 	}
 }
@@ -87,10 +91,10 @@ func serverMock() *httptest.Server {
 
 func notesMock(w http.ResponseWriter, _ *http.Request) {
 	mockNotes := NoteMailBody{
-		Notes:      nil,
-		EventStats: nil,
-		ImageUrl:   "",
-		Footer:     "test",
+		Notes:    nil,
+		Events:   []event.Event{},
+		ImageUrl: "",
+		Footer:   "test",
 	}
 	b, _ := json.Marshal(mockNotes)
 	_, _ = w.Write(b)
