@@ -1,7 +1,7 @@
 import {TestBed} from '@angular/core/testing';
 
 import {HildeGuard} from './hilde.guard';
-import {WebStorageModule} from 'ngx-web-storage';
+import {NgxWebstorageModule} from 'ngx-webstorage';
 import {LoggerTestingModule} from 'ngx-logger/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 
@@ -10,7 +10,7 @@ describe('AuthGuard', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-    imports: [WebStorageModule, LoggerTestingModule, RouterTestingModule],
+    imports: [NgxWebstorageModule.forRoot(), LoggerTestingModule, RouterTestingModule],
     teardown: { destroyAfterEach: false }
 });
     guard = TestBed.inject(HildeGuard);
