@@ -1,7 +1,8 @@
 import {Injectable} from '@angular/core';
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree} from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import {Observable} from 'rxjs';
-import {WebStorageService} from 'ngx-web-storage';
+// import {WebStorageService} from 'ngx-web-storage';
+import {LocalStorageService} from 'ngx-webstorage';
 import {NGXLogger} from 'ngx-logger';
 
 export const PRE_LOGIN_URL_SESSION_KEY = 'preloginUrl';
@@ -13,11 +14,11 @@ export const PRE_LOGIN_URL_SESSION_KEY = 'preloginUrl';
  * https://juristr.com/blog/2018/11/better-route-guard-redirects/
  * https://stackblitz.com/edit/angular-auth-guard-service?file=src%2Fapp%2Fauth.service.ts
  */
-export class HildeGuard implements CanActivate {
+export class HildeGuard  {
 
   private readonly className = 'HildeGuard';
 
-  constructor(private storage: WebStorageService,
+  constructor(private storage: LocalStorageService,
               private logger: NGXLogger,
               private router: Router) {
   }
@@ -25,10 +26,10 @@ export class HildeGuard implements CanActivate {
   canActivate(
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    const previousPath = this.storage.session.get(PRE_LOGIN_URL_SESSION_KEY);
+    const previousPath = this.storage.retrieve(PRE_LOGIN_URL_SESSION_KEY);
     if (previousPath) {
       this.logger.info(`${this.className}.canActivate: ${PRE_LOGIN_URL_SESSION_KEY} found in session, redirecting to ${previousPath}`);
-      this.storage.session.remove(PRE_LOGIN_URL_SESSION_KEY); // clean for next home access
+      this.storage.clear(PRE_LOGIN_URL_SESSION_KEY); // clean for next home access
       this.router.navigateByUrl(previousPath);
       return false;
     } else {
