@@ -2,9 +2,14 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/cloudevents/sdk-go/v2/event"
+	"github.com/segmentio/kafka-go"
+	"github.com/tillkuhn/rubin/pkg/polly"
 	"html/template"
+	"io"
 	"log"
 	"net/mail"
 	"os"
@@ -12,6 +17,10 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/tillkuhn/rubin/pkg/rubin"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 
 	"github.com/dustin/go-humanize"
 	"github.com/tillkuhn/angkor/go/topkapi"
@@ -36,10 +45,10 @@ type Note struct {
 
 // NoteMailBody wraps a list of Notes plus an optional footer to compose the outbound mail
 type NoteMailBody struct {
-	Notes      []Note
-	EventStats map[string]int
-	ImageUrl   string
-	Footer     string
+	Notes    []Note
+	Events   []event.Event
+	ImageUrl string
+	Footer   string
 }
 
 var (
@@ -51,6 +60,8 @@ var (
 	ReleaseName = "pura-vida"
 	AppId       = "remindabot"
 	logger      = log.New(os.Stdout, fmt.Sprintf("[%-10s] ", AppId), log.LstdFlags)
+	// kafka topic for app events
+	topic = "app.events"
 )
 
 // SSL/TLS Email Example, based on https://gist.github.com/chrisgillis/10888032
@@ -70,6 +81,25 @@ func main() {
 		logger.Fatalf("Error publish event to %s: %v", "system", err)
 	}
 
+	// Experiment with new Kafka Setup (rubin)
+	payload := map[string]string{"app": AppId, "version": AppVersion, "status": "started"}
+	rc := rubin.Must[*rubin.Client](rubin.NewClientFromEnv())
+	// fmt.Println(event)
+	if resp, err := rc.Produce(context.Background(), rubin.RecordRequest{
+		Topic:        topic,
+		Data:         payload,
+		AsCloudEvent: true,
+		Source:       "urn:app:angkor/" + AppId,
+		Type:         "net.timafe.event.app.started.v1",
+		Subject:      "reminder",
+		// Headers: map[string]string{}, // can be nil
+	}); err != nil {
+		logger.Printf("push message to %s failed: %v", topic, err) // but continue
+	} else {
+		logger.Printf("successfully pushed message to %s: %d", topic, resp.ErrorCode)
+	}
+
+	// let's get back to our actual business
 	reminderResponse, err := fetchReminders(config.ApiUrl, config.ApiToken, config.ApiTokenHeader)
 	// for whatsoever reason this might fail the first time with timeout, so we retry once
 	// err=Get ...: context deadline exceeded (Client.Timeout exceeded while awaiting headers)
@@ -81,7 +111,9 @@ func main() {
 		}
 	}
 
-	defer reminderResponse.Close()
+	defer func(reminderResponse io.ReadCloser) {
+		_ = reminderResponse.Close()
+	}(reminderResponse)
 	// Parse JSON body into a list of notes
 	var notes []Note // var notes []interface{} is now a concrete struct
 	err = json.NewDecoder(reminderResponse).Decode(&notes)
@@ -117,10 +149,28 @@ func main() {
 	}
 
 	// Pull recent events from topic to create a digest
-	actions := make(map[string]int)
-	kClient.Config.ConsumerTimeout = 5 * time.Second
-	consumeEvents(kClient, actions)
-	logger.Printf("Received actions %v", actions)
+	// actions := make(map[string]int)
+	// kClient.Config.ConsumerTimeout = 5 * time.Second
+	// consumeEvents(kClient, actions)
+	// logger.Printf("Received actions %v", actions)
+	p, err := polly.NewClientFromEnv()
+	logger.Printf("Starting polly client %v to check for new events", p)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
+	defer cancel()
+	var events []event.Event
+	topics := []string{"ci.events", "app.events", "system.events"}
+	err = p.Poll(ctx,
+		kafka.ReaderConfig{GroupTopics: topics, GroupID: config.KafkaGroupID},
+		func(_ context.Context, m kafka.Message) {
+			if ce, err := polly.AsCloudEvent(m); err == nil {
+				logger.Printf("Received event %s %s %s", ce.Source(), ce.Type(), ce.Subject())
+				events = append(events, ce)
+			} else {
+				logger.Printf("WARNING: Message %v cannot be parsed into an event", m)
+			}
+		})
+	logger.Printf("Received %d events", len(events))
+	p.WaitForClose()
 
 	// Prepare and send reminderMail
 	testFrom := "remindabot@" + os.Getenv("CERTBOT_DOMAIN_NAME")
@@ -128,10 +178,10 @@ func main() {
 	var buf bytes.Buffer
 	tmpl, _ := template.New("").Parse(mailTemplate())
 	noteMailBody := &NoteMailBody{
-		Notes:      notes,
-		EventStats: actions,
-		ImageUrl:   config.ImageUrl,
-		Footer:     mailFooter(),
+		Notes:    notes,
+		Events:   events,
+		ImageUrl: config.ImageUrl,
+		Footer:   mailFooter(),
 	}
 
 	// render the mail body
@@ -160,7 +210,8 @@ func mailSubject() string {
 }
 
 func mailFooter() string {
-	rel := strings.Title(strings.Replace(ReleaseName, "-", " ", -1))
+	cs := cases.Title(language.English)
+	rel := cs.String(strings.Replace(ReleaseName, "-", " ", -1))
 	year := time.Now().Year()
 	return "&#169; " + strconv.Itoa(year) + " · Powered by Remindabot · " + AppVersion + " " + rel
 }
