@@ -1,12 +1,12 @@
 package net.timafe.angkor.domain
 
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.Id
 import org.springframework.data.annotation.CreatedBy
 import org.springframework.data.annotation.CreatedDate
 import java.time.ZonedDateTime
 import java.util.*
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.Id
 
 @Entity
 data class Event(
