@@ -1,11 +1,12 @@
 package net.timafe.angkor.domain
 
-import io.hypersistence.utils.hibernate.type.basic.PostgreSQLEnumType
+// import io.hypersistence.utils.hibernate.type.basic.PostgreSQLEnumType
+import jakarta.persistence.*
 import net.timafe.angkor.config.annotations.EntityTypeInfo
 import net.timafe.angkor.domain.enums.LocationType
-import org.hibernate.annotations.Type
+import org.hibernate.annotations.JdbcType
+import org.hibernate.dialect.PostgreSQLEnumJdbcType
 import java.time.LocalDate
-import jakarta.persistence.*
 
 /**
  * Place 2 Go (Managed Domain Entity)
@@ -22,7 +23,8 @@ class Place(
 
     @Enumerated(EnumType.STRING)
     @Column(name = "ltype")
-    @Type(PostgreSQLEnumType::class)
+    // @Type(PostgreSQLEnumType::class)
+    @JdbcType(PostgreSQLEnumJdbcType::class)
     var locationType: LocationType = LocationType.PLACE,
 
     ) : LocatableEntity()
