@@ -1,9 +1,9 @@
 package net.timafe.angkor.domain
 
-import net.timafe.angkor.config.annotations.EntityTypeInfo
 import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
+import net.timafe.angkor.config.annotations.EntityTypeInfo
 
 /**
  * Entity that represents a Video
