@@ -1,7 +1,7 @@
 package net.timafe.angkor.repo
 
 import net.timafe.angkor.domain.Link
-import net.timafe.angkor.domain.enums.LinkMediaType
+import net.timafe.angkor.domain.enums.Media_Type
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
@@ -20,13 +20,13 @@ interface LinkRepository : CrudRepository<Link, UUID> {
 
     /* Feeds */
     @Cacheable(FEED_CACHE)
-    @Query("SELECT l FROM Link l where l.mediaType = net.timafe.angkor.domain.enums.LinkMediaType.FEED ORDER BY l.createdAt")
+    @Query("SELECT l FROM Link l where l.mediaType = net.timafe.angkor.domain.enums.Media_Type.FEED ORDER BY l.createdAt")
     fun findAllFeeds(): List<Link>
 
-    @Query("SELECT COUNT(l) FROM Link l where l.mediaType = net.timafe.angkor.domain.enums.LinkMediaType.FEED")
+    @Query("SELECT COUNT(l) FROM Link l where l.mediaType = net.timafe.angkor.domain.enums.Media_Type.FEED")
     fun feedCount(): Long
 
     // todo add index on name, append OrderByName
-    fun findByMediaType(mediaType: LinkMediaType): List<Link>
+    fun findByMediaType(mediaType: Media_Type): List<Link>
 
 }
