@@ -1,6 +1,7 @@
 package net.timafe.angkor.service
 
 import jakarta.annotation.PostConstruct
+import jakarta.mail.internet.MimeMessage
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.mail.MailProperties
@@ -8,7 +9,6 @@ import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.mail.javamail.MimeMessageHelper
 import org.springframework.mail.javamail.MimeMessagePreparator
 import org.springframework.stereotype.Service
-import jakarta.mail.internet.MimeMessage
 
 /** Mail Service see https://www.baeldung.com/spring-email */
 @Service
