@@ -1,12 +1,12 @@
 package net.timafe.angkor.web
 
+import jakarta.validation.Valid
 import net.timafe.angkor.config.Constants
 import net.timafe.angkor.domain.dto.LocationPOI
 import net.timafe.angkor.domain.dto.LocationSummary
 import net.timafe.angkor.domain.dto.SearchRequest
 import net.timafe.angkor.service.LocationSearchService
 import org.springframework.web.bind.annotation.*
-import jakarta.validation.Valid
 
 
 @RestController
