@@ -172,7 +172,7 @@ class IntegrationTests(
     fun testFeeds() {
         val items = linkController.getFeeds()
         assertThat(items.size).isGreaterThan(0)
-        assertThat(items[0].mediaType).isEqualTo(LinkMediaType.FEED)
+        assertThat(items[0].mediaType).isEqualTo(Media_Type.FEED)
         val id = items[0].id
         assertThat(linkController.getFeed(id!!).items.size).isGreaterThan(0)
     }
